requirements.txt
absl-py==2.1.0
accelerate==1.3.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anthropic==0.37.1
antlr4-python3-runtime==4.13.2
anyio==4.6.2.post1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
av==13.1.0
backcall==0.2.0
bc==0.0.1
beautifulsoup4==4.12.3
binpacking==1.5.2
bitsandbytes==0.45.2
black==24.10.0
blake3==1.0.4
bleach==6.2.0
blinker==1.8.2
blis==0.7.11
cachetools==5.2.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
colorlog==6.9.0
compressed-tensors==0.9.1
confection==0.1.5
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.11
dacite==1.8.1
DataProperty==1.1.0
datasets==3.2.0
debugpy==1.8.9
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.4
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
easydict==1.13
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.21.1
ffmpy==0.4.0
filelock==3.16.1
flake8==7.1.1
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.0
future==1.0.0
gekko==1.2.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
GPUtil==1.4.0
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.67.0
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
html5tagger==1.3.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.29.2
humanfriendly==10.0
identify==2.6.5
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
ipdb==0.13.13
ipython==8.12.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.10.0
joblib==1.4.2
jq==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
latex2sympy2_extended==1.0.6
libclang==18.1.1
liger_kernel==0.5.2
linkify-it-py==2.0.3
litellm==1.51.0
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lmdb==1.5.1
loguru==0.7.3
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
math-verify==0.5.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
mistune==3.0.2
modelscope==1.18.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.14.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
open_clip_torch==2.28.0
openai==1.52.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
optimum==1.23.1
orjson==3.10.7
oss2==2.19.0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.12.0
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pre_commit==4.0.1
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.0
pytablewriter==1.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.9
sacrebleu==2.5.1
safetensors==0.4.5
sanic==24.6.0
sanic-routing==23.12.0
scikit-learn==1.6.1
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.21.0
setproctitle==1.3.4
sglang==0.3.4.post2
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
srsly==2.5.1
stack-data==0.6.3
starlette==0.40.0
sty==1.0.6
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
textual==1.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.0.2
tomlkit==0.12.0
torch==2.5.1
torchao==0.6.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.5
tracerite==1.1.1
traitlets==5.14.3
transformers==4.49.0
transformers-stream-generator==0.0.5
triton==3.1.0
trl==0.14.0
typepy==1.3.4
typer==0.12
typing_extensions==4.12.2
tyro==0.8.12
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
virtualenv==20.28.1
vllm==0.7.2
wandb==0.18.3
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.4
xformers==0.0.28.post3
xgrammar==0.1.11
XlsxWriter==3.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.15.3
zipp==3.20.2
zmq==0.0.0